fix: change primary key type from Integer to Long in NoteRepository interface

+++ src/main/java/com/mindvault/backend/repository/NoteRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface NoteRepository extends JpaRepository<Note, Integer> {
+public interface NoteRepository extends JpaRepository<Note, Long> {
     List<Note> findByUser(User user);
 }
